Add user authentication, sessions, and flashing
Things are moving along. Users can now log in and a session will be
created. The user is being given feedback about what the application is doing.
Have to decide on what to serve to the user. Shall I add logic to the template
or just render to different templates based on if the user is logged in.